nội dung

+++ code56.cpp
@@ -6,6 +6,6 @@ using namespace std;
 int main() {
     int a;
     cin >> a;
-    cout << pow( a , 2) <<endl;
+    cout << (int)pow( a , 2) <<endl;
     return 0;
 }
